final modifications

+++ src/main/java/com/perez/model/OrderItem.java
@@ -20,7 +20,8 @@ public class OrderItem {
     private Food food;
 
     private int quantity;
-    private Long totalPrice;
+    //private Long totalPrice;
+    private double totalPrice;
 
     private List<String> ingredients;
 }
